Support worn armors.
Copying these correctly is a challenge due to the way armors and addons branch out into other parts of the record structure. Here we make several assumptions about things that are "not important" in the context of Worn Armors specifically.

Races also need to be kept consistent - while it's OK to specify DefaultRace for the armor itself, the addons need to actually match the NPC race. Current implementation spams all the races that could possibly make sense into the Valid Races list, which seems to be working for all the major mods, like Bijin. There's a risk of using more races than is really necessary, from more plugins than is really necessary - i.e. can include random custom races from anywhere in the load order - but this is not really causing any specific problems and won't add masters that weren't already there. If necessary, it can probably eventually be optimized to use a second pass after all the face imports are done, and prune the valid races to those NPCs that actually need it.

Fixes #7