Avoid copying vanilla overrides into output.
The new build system is aggressive in copying all dependent resources, but we probably don't want to include things like whatever texture pack is currently overriding the vanilla `femalehead.dds` and such.

If a texture exists in the vanilla archives (other than face tints), don't copy it. This logic is only applied to textures and not meshes, at least for now. There's an outside chance that swapping vanilla body replacers mid-game could lead to a conflict because of this, but if that turns out to be a serious issue then a similar filter can be added for meshes.

The implementation is achieved by a filter abstraction, so that services not normally included in the build task can be incorporated into the rules.