Fix various deficiencies in wig detection.
The dewiggifier now potentially conflicts with body replacer merges, since they both use Worn Armors. To minimize conflicts, the wig analysis is now much more careful about when to report a wig, and will only do so when it is clear that the Worn Armor contains nothing else of interest - i.e. when every other AA is either already part of the default skin or does not apply to the specific NPC.

Consequently, Dewiggify Records now removes the Worn Armor, which was previously omitted.