Fix potential block/stall bug in build pipeline.
The previous code, by directly invoking the task-creation method, could end up synchronously running a method despite it being declared `async`. The only way to guarantee that the task does not block the pipeline is to use `Task.Run` or equivalent.

The reason this is important is that the task cannot be added to `remainingTasks` until we actually have a `Task` instance, and if it is not added in time, then when another running task completes, the pipeline can appear stuck and trigger the corresponding failure.

One other minor change is to lock the results dictionary - which is accessed by multiple threads - and ensure that task continuations run on the pipeline's main thread so that the other state does not need to be synchronized.