Restore the de-wiggifier.
Implementation is very similar to the original, with a few modifications to fit into the build task framework. First, now that tasks can run concurrently, it needs a bit of synchronization to avoid conflicting with the texture scan (via the FileSync). Also, the Nifly code has been updated to open the file asynchronously and load it using the stream... unfortunately, _saving_ cannot be done asynchronously or via stream, since the SWIG bindings don't currently expose any way to do it.

Aside from this it's looking good, both NifSkope and in-game show the desired appearances (with hair).