Fix scrolling of build tasks.
Too many tasks would be cut off at the bottom of the screen with no way to scroll down (other than stretching the window size).

Uses a hack to "suppress" mouse wheel events on the build data grid, which normally tries to trap them even if the grid can't scroll. The scroll viewer is instead placed on the parent control, so that it doesn't interfere with the prior margin hack on the grid.

Slightly increases the default window height - 720 being the actual height of the lowest conceivable modern screen resolution, however unlikely it is.