Add analyzer support for skin inheritance.
Similar to head parts, NPCs without a custom Worn Armor are actually just inheriting it from the race (usually `SkinNaked`). Updates the logic to incorporate this into change detection and provide the "effective skin" as an output. Also changes the wording of "body" to "skin" to be consistent with game terminology.

This is important for the case of preserving the appearance provided by a race edit in an NPC overhaul without actually accepting the new race. The builder will need to know which skin the NPC should have in order to line it up with face textures.

Extends #7