Handle removed warning types on load.
These may still be present in the app settings despite being removed from the original enum. To avoid having to keep them around forever, use a "safe" JSON item converter and forward all obsolete warnings to a default value, which is removed on load and hidden from the UI.

Also fixes plugins not being displayed in the settings dropdown - these were simply not populated, forgotten in the big migration.