Fix a possible bug adding new NPCs to profile.
When restoring autosaves, it was possible for new NPCs to never be added to the autosave afterward, due to the policy defaults not actually making any real changes (e.g. having only 2 options where both the pre-policy and post-policy defaults pointed to the 2nd one).

New NPCs will now always be logged to the autosave. The implementation should prevent redundant entries, as the "force write" only occurs when explicitly requested and when no implicit event was generated.