Buffer initial log events for view model sink.
Early events - mainly the version info - would get written to the log file, but would not show up in the log window due to late assignment of the view model to the sink.

The sink will now keep a buffer of log events until a view model is assigned to it.