Make MO2 mod keys always have names.
Earlier logic was to use a default name (component name) _only_ if neither the ID nor name was known. But not every caller is going to look up the mod and find a default component if the key doesn't have a mod name.

Instead, we should use the component name as a default mod name, which is how it already works in the Vortex code, and also how it works if we can't _find_ the MO2 information at all - the "fully missing" path should not be more reliable than the "partially missing" path.