Include NPCs with FaceGen overrides in profile.
Ensures that standalone facegens are counted toward both startup and profile filters when checking for at least one, or more than one face override.

Since all standalone facegens indexed from the mod directory are an alternative option, just one is enough for inclusion. And since facegen overrides are never picked by default, these should be shown even if there is only one facegen override, even when filtering to "multiple" choices.

To avoid some of the confusion this may cause, the mugshot viewer and accompanying data now reports if a mod is disabled, aside from simply saying "plugin not loaded". This isn't perfect, but helps at least somewhat in determining if a facegen override is actually a facegen override as opposed to a missing or unloaded plugin.

Fixes #101