Add changelog bump script, add to release workflow.
Should now automate all the important parts of the release - rolling up the project and changelog versions, building the release package, creating the release, and committing the changes.

There does not seem to be a way to atomically do the release and commit/push in a single step, so commit/push is saved for the last step, since a release can be pulled without too much difficulty while a commit is more complicated to "undo".