Add simple BAML-JSON conversion.
This is a convention-based and fairly limited system that understands the "paragraph numbering" being used in the example target. These will be "reduced" from their UIDs to simple index placeholders - with the index being the actual index in the `Messages` array, not the original index implied by UID which may include non-localizable elements.

Placeholder replacements are also supported through localization comments. Since it's impossible to get these to come through the `BamlLocalizer` (it will totally ignore elements like `Viewbox`, no way to read the comments), we **must** have these compiled to a standalone localization file.

Non-placeholder references which aren't numbered can also be shortened to their "local ID", if the full UID is an extension of the segment UID.

This is far from perfect, and still needs a lot more work, particularly in the areas of round-tripping (may need to save reverse-transformation info to a separate file for use when merging), and identifying *new* non-localized messages or changed messages instead of processing the whole assembly.