Add item logging to build pipeline.
This was intended to be in the original implementation and accidentally left out. When in debug mode, the build pipeline should log each individual item being funneled to the UI, same as the old build progress view model did.

Also adds thread ID to the log format so that it is easier to interpret logs of parallel processes. This is most important for debug logging so that the last debug log can be connected to an exception, if there was one. Without this, it can be hard to know which item caused the exception.