Add reference-checking utility.
Can be configured with expressions pointing to all or most common record expansions, and will follow all paths and report the full paths from the root record to any unresolvable references.

This should help to produce useful, user-friendly (to the extent that is possible) output about plugins that have these problems and exactly why they are problems, as well as pointing directly to the records (i.e. NPCs) that need to be suppressed.

Still needs to be incorporated into the other analyzers.

#107