Disable and warn on invalid references.
When NPCs in plugins dead-end at some broken form link, they will be shown in an error report at startup for prominence, and subsequently disabled in all selection functions and marked disabled in the UI. This will prevent what would otherwise appear to be either a crash on build (current behavior) or some undefined and probably wrong behavior (if we ignored them).

Considering an "ignore permanently" but not sure if the benefits outweigh the potential risk of users just clicking anything they can to get rid of the errors without paying attention.

Still need doc page which is linked by the warning, will close after that's done.

#107