Various dark-mode fixes.
Mostly on the Profile page, but elsewhere have tried to use system colors instead of custom colors where possible. For example, "danger" color is slightly different from system "error" color in light mode (red vs. slightly darker red), but dark mode uses a completely different color (yellow).

There is no perfect fix for the profile icons in dark mode, because anything that is exceptionally clear under normal circumstances becomes almost invisible when highlighted. Only neutral colors are as legible with and without the highlight, but that defeats the purpose of using color as an additional signal. Nevertheless, these colors should be much easier to see most of the time.

Fixes #100