Auto-switch mod directory to mod manager setting.
This is a togglable feature that is opted-in by default, as the "persistence" of the mod directory setting has been a point of confusion.

When enabled, if the app is launched from a mod manager, it will automatically index and use the mod directory where the manager points to, instead of the user setting which is now renamed to "default".

Combined with proper detection of MO2 forced portable instances, this should rarely make any mistakes, and if it does, the feature can be disabled in the settings menu.