Stop Mutagen environment from loading too early.
This regressed in a previous version and was never detected. It was the reason why the app seemed to be taking longer to get to the first pixel, but spent no time at all creating the load order analyzer. The `SettingsViewModel` was depending on the `IGameSettings` which in turn caused the environment to be created. Effectively, this completely ignored any customizations being made to the plugin selections at startup.

It's too valuable to have global access to these types to consider moving them all into providers and factories and such, and doesn't really help that much anyway. Using `Lazy<T>` can alleviate the issue somewhat by deferring access to until the load order list is needed - but unfortunately this isn't quite good enough because starting the app in first-time mode will trigger this right away.

Instead, we add a "future"-like type, that's a little more intelligent than `Lazy<T>` and uses that under the hood, but requires a bit more wire-up and post-activation logic from the IoC container. Other types can now accept an `IAfterSetup<T>`, which will never actually cause the value to be resolved, instead will throw if the dependency has yet to be created - essentially a "weak reference" version of the Lazy type, with a callback version that is designed to work with property-change notifications.

Fixes #113