Read game path from Mod Organizer INI.
This will pick an accurate path much more often than GameFinder _in the case when EasyNPC is run from a mod manager_, which is going to be most of the time. Similarly to #111 and auto-switching to the mod manager's mod directory, we pretty much always want the two settings to be in sync.

Like the profile, this setting can either be a plain string or the odd wrapped `@ByteArray(string)`.

Now also logs the selected game path for debugging/troubleshooting.

Closes #114