Refactor main window contents into view.
This seems like the least messy way of providing the ability to host other "app modes" in the same window, particularly those that don't require or want the nav view or loader logic.

Like the `MainView` itself, the `MainWindow` will now take its `DataContext` externally and display a template based on the actual top-level view model, with the current/default obviously being the extracted `MainView`. The post-build report along with its loader/log will be a separate top-level model.

This also allows us to reuse the IoC container, error-handling, etc., and we'll want all of that for the post-build report.

#78