Add preliminary NPC checks to post-build.
These are producing too many false positives right now, but the basic idea is there and the performance is finally up to scratch (no more than a few seconds for a reasonably small number of conflicts).

Updates the `ArchiveIndex` to order results by the actual archive order, which is necessary to get equivalent behavior to the `GameFileProvider` (but with much better performance). This may cost a bit of performance in other parts of the app - not yet tested.