Support multiple "additional" head parts in analysis.
Marks/scars (and possibly other parts?) can be layered and do not need to have a single canonical head part per type. Although usage of this seems rare, the post-build checks were flagging the NPC merge for a handful of NPCs because it only had one mark while the facegen itself (and the original NPC data) had two.

If multiple marks are found, use them all.

Also removes the "extra part" check from the main head parts checker. It doesn't seem like there are any mods that actually include _unnecessary_ extra parts (i.e. those implied by a main part) in the record data. However, there _are_ mods that uses extra parts _not_ included in any main part, especially marks/scars, and this is allowed. Unsure if the correct logic is to remove it entirely or to only filter if it's an extra part **and not also** a secondary part, but the current implementation seems to have no issues.

#78