Implement archive-extraction post-build workaround.
Presents as an interstitial-like panel and has basic task progress/item tracking as files are copied.

There is no way to prevent having to restart the app in order to see the latest updates, since MO2 needs to update its VFS and Vortex needs to deploy. The task background will change when extraction is complete, which should be OK feedback for now.

Copies both facegens and facetints, although explicit face tint consistency checking is still not implemented yet.

#78