Overhaul the pre-build scroll/animation code.
Should now "monitor" the scroll position (update the selected list item), save the current position when navigating to another tab, and be a bit smoother in general.

Rules are kludgey and heavy on code-behind, but there doesn't seem to be a great alternative considering how we need to deal directly with the UI elements.