Exclude head part meshes/morphs from missing assets.
These are not completely necessary with NPC overhauls and many mod authors deliberately exclude them. "In theory" the game only looks at the compiled facegen. There are probably some breaking edge cases, but for now there's no need to scare users with alerts that probably don't matter.

This is implemented by collecting the actual record sources of asset references. This seemed better than trying to shoehorn application-specific logic into a general-purpose analyzer, and safer than excluding the paths in that it still tracks those assets and allows them to be copied if available. The only change to the analyzer itself is a `Link` function that allows each source record a chance to mutate the accumulator value, even if the value itself was cached.