Add behavior for restoring scroll position.
The need for this is very common due to how the navigation view doesn't keep pages loaded (unfortunately). For the build preview especially, since there's an expectation of jumping back and forth between profile and preview, it's important to preserve scroll position. This aims to make it as easy as possible to do so.