Remove Playable flag from imported head parts.
These tend to crash the race menu when selected. Even if the head part is originally from one that is designed to be playable (like High Poly Head), the imported version is only used by NPCs and therefore should still not be playable, otherwise would duplicate the playable one.

Fixes #124