Patch Head Part's Valid Races to use merged keys.
Works (almost) the same way as it does for Worn Armor addons. As NPCs are imported, their races are registered into the importer which adds them to the form list. Ensures that the form list contains all of the races that are actually used by merged NPCs, and no other races.

Fixes #119