Allow vanilla mugshots to be double-clicked.
Uses a separate method to revert to vanilla, which is called when the mugshot is marked as being base-game related.

Also fixes the bug where double-clicking on a mugshot without a facegen wouldn't work. This was caused by the FaceGen Redirect fixes a while back. Allowing these faces to be chosen as long as there is a corresponding plugin (even if there is no facegen) should hopefully fix the bug without breaking the FG override logic.

Fixes #122