Fix overzealous template-NPC check.
Previous logic would check for "exactly one distinct trait parent among all options", which ignored options with **no** template. So in rare cases when the NPC is normally not a template at all, but one plugin decides to make it a template, this would be seen as "just one template".

Now correctly implemented as "exactly one distinct trait parent **used by all** options", i.e. any options without a template or without trait inheritance will short-circuit this check to false (i.e. not a single template).

Also updates to do this computation once, when the NPC model is created, because it's not free to check this repeatedly.

Fixes #118