Always import injected records.
"Injection" can trick the importer into thinking the record doesn't need to be imported because it's in a plugin that's going to be a real master (like 3DNPC), but where the record will actually disappear when some overhaul (INVO) is disabled.

This is a quick-and-dirty workaround that forces every injected record to be imported. This may result in more records than necessary (e.g. if the injecting plugin is _also_ a master), but as there is no way to actually trace the "true" source of such a record, no solution is likely to be perfect and even a marginally better solution could be very complicated and unpredictable.

There's a slight performance hit as thousands more records need to be looked up, but in practice it doesn't make a huge difference - at most a few seconds in one build task, while others are running in parallel.

Fixes #95