Add game data path from Vortex extension.
This had been added to the EasyNPC side of the interface, but was never added to the extension to populate it. Now obtains this from the only place where this path seems to exist in the Vortex state, the `discovered` section of the Vortex settings.

Also updates EasyNPC Vortex code to read this value - was previously hardcoded to an empty string.

Closes #121