Fix the Post-Build report for Vortex.
There were two main issues here. First is that the ordering of search results was totally random, causing different mods to be the "actual facegen source" each time. Second was that if the source different, _even if it had the correct/identical file_, some head part comparisons were failing because of improperly converted SSE mods (Bijin!) still using the legacy `NiTriShape` instead of `BSDynamicTriShape`.

In the latter case, we don't actually care which one it is, so the code has been updated to handle either type.

The former is now handled by processing Vortex's (undocumented) file prioritization state to pick the winners, coded to fail silently if Nexus changes the way this works.

There were never any actual conflicts - as some of the users reported, the BSA extraction workaround actually was successful. The post-build report was just giving false positives.

Fixes #123