Limit to used archives in Bad Archives check.
If an archive in some totally unrelated mod has issues, that's not very interesting. This should only matter if an archive belonging to one of the NPC mods we are trying to merge is corrupted. This can qualify as a blocking error, as it is currently defined, as opposed to a warning.

While it is possible that some unrelated archive corruption in an upstream mod is causing problems (e.g. broken High Poly Head or KS Hairdos), it's impossible to predict how an unrelated archive might affect the merge, and hopefully this would show up as a missing asset if it's important.

This also adds the first unit test for EasyNPC main app because it's actually easier to unit test this than to manually corrupt and test a bad archive, especially with the dual-BSA logic.