Update obsolete missing-mod-directory help text.
The "patch-only" mode never _actually_ worked, and although it was initially considered as a "maybe" feature in early development, it no longer makes any sense at all with the current design and would be a hard pass even if requested.

Get rid of the verbose and misleading warnings and stick to the facts.