Handle empty sequences in Alerts VM.
`Enumerable.Max` throws on empty sequences. We can use `DefaultIfEmpty` to ensure a non-empty sequence.

Also updates all the off-thread subscriptions in the Build Preview models to catch and log their errors, as these should all technically be non-blocking. If the app is going to fail anyway, might as well fail in the direction of letting the user continue with the build (and if the build itself crashes, they'll get a message for that).

It's not ideal - should probably show a clear warning when the preview data is known to be incorrect due to an exception - but good enough for a hotfix.

Fixes #131