Patch worn armors to include beast transformation addons.
Mechanism right now is to have a configurable (but hard to discover, thus hard to screw up, but technically modifiable in case of emergency) list of transformation races, and will scan and include all the addons for all the default skins for those races.

This means every used worn armor including the vanilla ones must be imported into the merge and patched, but vanilla (or master-contained) records can be only shallow-copied, not deep-copied.

Fixes #133