Bind EBO on every VAO bind.
The Silk.NET examples are wrong or misleading here; EBO doesn't really bind to the VAO, it binds directly to the GL context. It may make sense to logically associate one with the VAO, if the VAO represents a mesh or group of them, but the VAO wrapper must then be responsible for rebinding the EBO.

This fixes bugs related to geometry either not showing up (if the initial indices are equivalent to vertex order and the extra geometry goes at the end), or being generally corrupted (if interleaving happens very early on). Prior commits were effectively just rendering the original Buffer as-is and ignoring the EBO entirely.