Fix unsafe tangent/bitangent computations.
The popular algorithm doesn't take into account multiple vertices with the same UV coordinate. Under these circumstances, the r-value goes to infinity and tangents/bitangents also become infinity (or NaN, after orthogonalization).

A workaround for the same problem is presented at https://github.com/assimp/assimp/issues/230, although the second part of the logic isn't very clear - there doesn't appear to be any situation that could lead to a valid tangent but invalid bitangent, or vice versa - so that part of the logic is not implemented, only resetting the UV deltas when `r` is invalid.

Also completes the orthogonalization so that bitangents are also made orthogonal to both the normals and tangents. Doesn't seem to make a huge difference, but should technically give smoother lighting.