Add a posing/rigging abstraction.
Loosely modeled after how it's done in the NIF format, but hopefully less confusing. A skinned mesh is set up with per-bone inverse transforms (i.e. how to get from the skinning point to the rest or "bind" position) which can then be multiplied by the actual bone transform in scene.

Long-term, it might be better to export the transforms and let the shaders do the posing; this is much better for animation, but for now, since there won't be any animation and the goal is just to load a single base pose from the skeleton, having it abstracted this way is a little cleaner.