Add Scene data types and a new renderer for scenes.
Scene is, in this context, just an aggregation of objects which have meshes and texture sets. They will probably need to get updated later to include other shader settings as well - specular strength/color, alpha, shininess, all the stuff that's in the NIF shader property.

The scene renderer is designed to load both geometry and textures in the background, but this requires synchronization with OpenGL. Various other attempts to make this work via `MakeCurrent`/`MakeContextCurrent` have failed (GL always reports that it is "in use" even if we try to clear the context ASAP), and Silk's `Invoke` methods just throw a `NullReferenceException` for incomprehensible and un-debuggable reasons. The workaround right now is to use a custom scheduler, which is pretty crude and might have bugs.