Add a customizable file provider.
`GameFileProvider` is the canonical version but sometimes, especially in standalone or test apps, it helps to load specific paths and/or BSAs. This is also how tools like NifSkope and Outfit Studio behave, they look at the environment to see what's available but allow the user to pick specific BSAs in order to improve performance. We don't necessarily want to search every BSA.