Update NIF loader/renderer to pull textures (asynchronously) from file provider abstraction.
This allows them to read from BSAs, or any other source. Performance is about the same as before.

Adds a couple of explicit disposals of NIF objects, although they are probably not necessary. The rendering test app seems to use a lot of memory so this is making sure there are no leaks in this area.