Support for several Bethesda-specific shader options.
All of these options are now configurable via the `SceneObject` and include many of the more basic settings like lighting strengths/colors, glossiness (shininess, etc.).

Specular maps and object-space normal maps are now also supported. The latter required a fix to the shader; tangent-space normals can do without the inverse normal transform but model-space needs it and is totally broken without it.

Shaders now also fall back to straight normals, if available, when the normal map texture is not available.

Fixed the object bounds calculation, which was improperly handling meshes that didn't intersect the world origin. This was causing headaches with model transforms, trying to center the objects.