Use BCnEncoder for DDS, with fallback.
This library is much faster than the DirectXTex interop that was adapted from PDN. However, it does seem to fail in a few instances, i.e. divide-by-zero errors on the normal maps for body/hands/etc. textures in Diamond Skin. This requires a fallback to the native version for those textures, at least for now, until some kind of fix presents itself either here or in the library.

Memory usage initially appears to be much higher, but almost all of it gets garbage-collected and the total ends up being much lower. The unpredictability of this isn't the best, as the big GC passes will cause missed frames, but on the other hand, this isn't intended for a whole game engine, just to do some previews.