Add swizzle setting to scene objects and renderer.
Model-space normals behave strangely, most of the time the lighting seems correct when swapping the B and G channels (same as the Blender NIF importer) although for other maps it may not be right. Either way, allowing this to be specified at the model level and passed down to the fragment shader should be the way to go, and give the best chance of fixing all the little issues that come up. More swizzle settings may need to be added.