Add a first level of texture caching, using LRU cache.
This is a temporary half-measure and only really makes a difference for preloaded DDSes - although that is technically all the textures being used right now. However, it also only caches the "input". The "output" still needs to recomputed (copied) every time.

It should be possible to apply very similar logic in the OpenGL code and cache the `Texture` objects directly, which would be ideal. The current system helps prevent loading the same textures multiple times, but it would be better to also only marshal and bind them once each.