Support cubemaps/environment textures.
Uses reflection map, or constant value of 1. Looks like other tools might use the specular value, we don't do that yet.

Cubemaps only supported by BCnEncoder source, not implemented in DirectXTex version yet, may not be necessary if all cubemaps can be loaded by BCn.

Includes some minor housekeeping: replace flags in shaders with actual bools (still need to be set in the calling code as int). Fixes yet another mistaken/misleading assumption inherited from the Silk examples, implying that it is possible to completely decouple the active texture slot from the bound texture; in reality it is not that simple, and although it sort of works with only 2D textures (though RenderDoc shows some oddities with the slots), it breaks entirely when cube textures are added because OpenGL tries to interpret slot 0 (diffuse) as a cube.

Cube maps have to be set up with the ACTIVE UNIT (not just the bound slot) to where the cube map will end up. It's best to just require the slot/unit to be specified when creating the texture in the first place.