Add a Scene VIewer helper.
Provides an interactive view in which the mouse can be used to look around. Wheel/right-button zooms, left button rotates, middle button pans. Rotation is a bit buggy because the Y rotation is global rather than local to the current view - will fix sometime later.

Updates the Grid Renderer to support different planes. Since NIFs generally need to be rotated 90° around the X axis in order to have the correct orientation in OpenGL, the grid renderer should actually be set up in the Z plane rather than Y plane, if the grid is going to rotate, which is the desired effect here. (Also enables model rotation on the grid, since we want the mouse to rotate the model, not the entire view including lights).