Allow a default pose to be specified for NIF scenes.
Without GPU skinning, this is currently the simplest way to apply a pose that's not in the model file itself (i.e. in a separate skeleton). It will still use the pose found in the model file if a separate one is not specified.

It's a little hard to "reach" the mesh from the renderer tree from the outside world, so passing it to the constructor is convenient, albeit read-only.

Also fixes skin data not being applied for shapes using (legacy?) `NiSkinInstance` instead of `BSDismemberSkinInstance`.