Support alpha blending and alpha test.
Required for most head parts like hair, eyes, brows, etc. Alpha testing (i.e. `glAlphaFunc`) is no longer supported in OpenGL, so the fragment shader has to do it.

Makes a few other changes/fixes to the fragment shader to bring the lighting somewhat more in line with the front-runners, including tone mapping, falling back to specular for environment strength, and changing the order of color mixing. The result is still not perfect but it's looking like a fair approximation.

Note: Ideally the renderer should draw transparent meshes according to their distance from the camera, in order to get correct results. In practice this doesn't seem to be necessary for facegens, or other "basic" meshes, but could become an issue later.