Throttle texture path extraction and add timeouts.
The throttling code is hacky stuff that should be easily replaceable by `Parallel.ForEachAsync` in .NET 6, and probably be more stable with that API, but we don't have it in .NET 5.

A 10-second timeout per file is also added, which might help identify culprit files if there really is some kind of hardware/disk issue going on for the troubled users.

Fixes #165.