Overhaul of file-related build error handling.
Instead of just ignoring missing/uncopyable files, these are now reported to a metadata file stored in the mod output directory, along with some other basic info. The goal is to eventually use this for other features like build continuation, incremental builds, etc. Also possibly incorporation into post-build checking.

Updates the timeouts for texture path extraction to default to 30 seconds, to be configurable in the build settings UI (can turn off completely by setting to 0), and to log an error/report to the build metadata instead of "crashing" the build entirely.

#165